// src/app/Medical/Models/TourSignatures.ts
export class TourSignatureMapData {
  ID: any;
  TOUR_ID: any;
  TOUR_PART_B_SIGNATURE_1: any;
  TOUR_PART_B_SIGNATURE_2: any;
  TOUR_PART_B_SIGNATURE_3: any;
  TOUR_CHECKLIST_SIGNATURE_1: any;
  TOUR_PLACE_ORDER_SIGNATURE_1: any;
  TOUR_PLACE_ORDER_SIGNATURE_2: any;
  TOUR_PLACE_ORDER_SIGNATURE_3: any;
  ADV_TOUR_ADV_BILL_SIGNATURE_1: any;
  ADV_TOUR_ADV_BILL_SIGNATURE_2: any;
  ADV_TOUR_GAR_SIGNATURE_1: any;
  ADV_TOUR_GAR_SIGNATURE_2: any;
  ADV_TOUR_MANJURI_ADESH_SIGNATURE_1: any;
  ADV_TOUR_MANJURI_ADESH_SIGNATURE_2: any;
}